clean up metrics and add lots of them
