+++ NuGet.config
@@ -2,8 +2,6 @@
 <configuration>
     <packageSources>
         <clear />
-        <add key="AspNetCore" value="https://dotnet.myget.org/F/dotnet-2-0-0-preview2-final/api/v3/index.json" />
         <add key="nuget.org" value="https://api.nuget.org/v3/index.json"/>
-        <add key="Octokit-CI" value="https://ci.appveyor.com/nuget/octokit-net" />
     </packageSources>
 </configuration>
