+++ src/Hubbup.Web/Diagnostics/Metrics/ApplicationInsightsMetricsBuilderExtensions.cs
@@ -1,6 +1,6 @@
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Hubbup.Web.Diagnostics
+namespace Hubbup.Web.Diagnostics.Metrics
 {
     public static class ApplicationInsightsMetricsBuilderExtensions
     {
