+++ src/Hubbup.Web/Diagnostics/Metrics/IMetricsBuilder.cs
@@ -1,6 +1,6 @@
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Hubbup.Web.Diagnostics
+namespace Hubbup.Web.Diagnostics.Metrics
 {
     // Marker interface to hang .Add[MetricSinkName] extensions off of.
     public interface IMetricsBuilder
