+++ src/Hubbup.Web/Diagnostics/Metrics/MetricsService.cs
@@ -4,7 +4,7 @@ using System.Threading;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
-namespace Hubbup.Web.Diagnostics
+namespace Hubbup.Web.Diagnostics.Metrics
 {
     public class MetricsService : IMetricsService
     {
