--- src/Hubbup.Web/Diagnostics/RequestMetricsMiddlewareAppBuilderExtensions.cs
@@ -1,12 +1,0 @@
-using Hubbup.Web.Diagnostics;
-
-namespace Microsoft.AspNetCore.Builder
-{
-    public static class RequestMetricsMiddlewareAppBuilderExtensions
-    {
-        public static void UseRequestMetrics(this IApplicationBuilder applicationBuilder)
-        {
-            applicationBuilder.UseMiddleware<RequestMetricsMiddleware>();
-        }
-    }
-}
