+++ src/Hubbup.Web/Diagnostics/Telemetry/IRequestTelemetryListener.cs
@@ -1,0 +1,9 @@
+using Microsoft.AspNetCore.Http;
+
+namespace Hubbup.Web.Diagnostics.Telemetry
+{
+    public interface IRequestTelemetryListener
+    {
+        void AddProperty(HttpContext httpContext, string name, object value);
+    }
+}
