+++ src/Hubbup.Web/Pages/Index.cshtml
@@ -1,15 +1,22 @@
 @page "{repoSetName}"
+@using Microsoft.AspNetCore.Http
 @inject Microsoft.AspNetCore.Hosting.IHostingEnvironment HostingEnvironment
 @{
     var repoSet = RouteData.Values["repoSetName"];
     ViewData["PageTitlePrefix"] = repoSet;
     ViewData["CurrentGroup"] = repoSet;
+
+    HttpContext.AddTelemetryProperty("RepoSet", repoSet);
+    HttpContext.AddTelemetryProperty("RepoSetView", "Standup");
+
+    var gitHubName = HttpContext.User.Identity.Name;
+    HttpContext.AddTelemetryProperty("GitHubUser", gitHubName);
 }
 
 <div class="row issue-list-row">
 
     <div class="col-md-2 col-md-push-10">
-        <a asp-controller="IssueList" asp-action="Index" asp-route-repoSet="@repoSet">Go to triage view</a>
+        <a asp-controller="Triage" asp-action="Index" asp-route-repoSet="@repoSet">Go to triage view</a>
     </div>
 
     <div class="col-md-10 col-md-pull-2">
