+++ src/Hubbup.Web/Startup.cs
@@ -1,11 +1,11 @@
 using System;
-using System.Diagnostics;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Reflection;
 using System.Security.Claims;
 using Hubbup.Web.DataSources;
-using Hubbup.Web.Diagnostics;
+using Hubbup.Web.Diagnostics.Metrics;
+using Hubbup.Web.Diagnostics.Telemetry;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
@@ -51,53 +51,49 @@ namespace Hubbup.Web
 
             services.AddMemoryCache();
 
-            services.AddAuthentication(options =>
-            {
-                options.DefaultSignInScheme = CookieAuthenticationDefaults.AuthenticationScheme;
-                options.DefaultAuthenticateScheme = CookieAuthenticationDefaults.AuthenticationScheme;
-                options.DefaultChallengeScheme = "GitHub";
-            });
-
-            services.AddCookieAuthentication(options =>
-            {
-                options.LoginPath = new PathString("/signin");
-
-                // Work around https://github.com/aspnet/Security/issues/1231
-                options.CookieSameSite = SameSiteMode.None;
-            });
-
-            services.AddOAuthAuthentication("GitHub", options =>
-            {
-                options.CallbackPath = new PathString("/signin-github");
-                options.AuthorizationEndpoint = "https://github.com/login/oauth/authorize";
-                options.TokenEndpoint = "https://github.com/login/oauth/access_token";
-                options.UserInformationEndpoint = "https://api.github.com/user";
-                options.ClaimsIssuer = "GitHub";
-
-                options.ClientId = Configuration["GitHubClientId"];
-                options.ClientSecret = Configuration["GitHubClientSecret"];
-                options.Scope.Add("repo");
-                options.SaveTokens = true;
-
-                options.Events.OnCreatingTicket = async context =>
+            services
+                .AddAuthentication(options =>
                 {
-                    var request = new HttpRequestMessage(HttpMethod.Get, context.Options.UserInformationEndpoint);
-                    request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
-                    request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", context.AccessToken);
+                    options.DefaultSignInScheme = CookieAuthenticationDefaults.AuthenticationScheme;
+                    options.DefaultAuthenticateScheme = CookieAuthenticationDefaults.AuthenticationScheme;
+                    options.DefaultChallengeScheme = "GitHub";
+                })
+                .AddCookie(options =>
+                {
+                    options.LoginPath = new PathString("/signin");
+                })
+                .AddOAuth("GitHub", options =>
+                {
+                    options.CallbackPath = new PathString("/signin-github");
+                    options.AuthorizationEndpoint = "https://github.com/login/oauth/authorize";
+                    options.TokenEndpoint = "https://github.com/login/oauth/access_token";
+                    options.UserInformationEndpoint = "https://api.github.com/user";
+                    options.ClaimsIssuer = "GitHub";
 
-                    var response = await context.Backchannel.SendAsync(request, HttpCompletionOption.ResponseHeadersRead, context.HttpContext.RequestAborted);
-                    response.EnsureSuccessStatusCode();
+                    options.ClientId = Configuration["GitHubClientId"];
+                    options.ClientSecret = Configuration["GitHubClientSecret"];
+                    options.Scope.Add("repo");
+                    options.SaveTokens = true;
 
-                    var payload = JObject.Parse(await response.Content.ReadAsStringAsync());
+                    options.Events.OnCreatingTicket = async context =>
+                    {
+                        var request = new HttpRequestMessage(HttpMethod.Get, context.Options.UserInformationEndpoint);
+                        request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
+                        request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", context.AccessToken);
 
-                    // Add GitHub claims
-                    context.Identity.AddClaim(new Claim(ClaimTypes.NameIdentifier, payload.Value<string>("id"), context.Options.ClaimsIssuer));
-                    context.Identity.AddClaim(new Claim(ClaimTypes.Name, payload.Value<string>("login"), context.Options.ClaimsIssuer));
-                    //context.Identity.AddClaim(new Claim(ClaimTypes.Email, payload.Value<string>("email"), context.Options.ClaimsIssuer));
-                    //context.Identity.AddClaim(new Claim("urn:github:name", payload.Value<string>("name"), context.Options.ClaimsIssuer));
-                    context.Identity.AddClaim(new Claim("urn:github:url", payload.Value<string>("url"), context.Options.ClaimsIssuer));
-                };
-            });
+                        var response = await context.Backchannel.SendAsync(request, HttpCompletionOption.ResponseHeadersRead, context.HttpContext.RequestAborted);
+                        response.EnsureSuccessStatusCode();
+
+                        var payload = JObject.Parse(await response.Content.ReadAsStringAsync());
+
+                        // Add GitHub claims
+                        context.Identity.AddClaim(new Claim(ClaimTypes.NameIdentifier, payload.Value<string>("id"), context.Options.ClaimsIssuer));
+                        context.Identity.AddClaim(new Claim(ClaimTypes.Name, payload.Value<string>("login"), context.Options.ClaimsIssuer));
+                        //context.Identity.AddClaim(new Claim(ClaimTypes.Email, payload.Value<string>("email"), context.Options.ClaimsIssuer));
+                        //context.Identity.AddClaim(new Claim("urn:github:name", payload.Value<string>("name"), context.Options.ClaimsIssuer));
+                        context.Identity.AddClaim(new Claim("urn:github:url", payload.Value<string>("url"), context.Options.ClaimsIssuer));
+                    };
+                });
 
             services.AddMvc(options =>
             {
@@ -107,15 +103,13 @@ namespace Hubbup.Web
                 }
             });
 
-            services.AddMetrics(
-                options =>
+            services
+                .AddMetrics(options =>
                 {
                     options.FlushRate = TimeSpan.FromSeconds(5);
-                },
-                builder =>
-                {
-                    builder.AddApplicationInsights();
-                });
+                })
+                .AddApplicationInsights();
+            services.AddSingleton<IRequestTelemetryListener, ApplicationInsightsRequestTelemetryListener>();
         }
 
         public void Configure(IApplicationBuilder app, ILoggerFactory loggerFactory, IDataSource dataSource, IApplicationLifetime lifetime)
@@ -127,7 +121,7 @@ namespace Hubbup.Web
             dataSource.ReloadAsync(lifetime.ApplicationStopping).Wait();
             logger.LogDebug("Loaded repo set and person set data");
 
-            app.UseRequestMetrics();
+            app.UseDiagnostics();
 
             if (HostingEnvironment.IsDevelopment())
             {
